simulation_almost_perfect
now the simulation runs a lot faster (less then 1 min, before it runs in 1 hour). the output file was refformated and now it has native matlab support. You only have to open the file and run it on matlab workspace. The file location is ..project_34_segment.sim\sim_1\behav\resultados.m
now we have a lot to work on the main file: add  apellidos, identificacion and grupo; add color support and a lot of bug fixes. Simulating it will be easy!